fix: run knowledge ingest using python gptscript sdk. Rename folder to file-reader
