+++ knowledge/file-loader.gpt
@@ -7,4 +7,4 @@ Credential: github.com/gptscript-ai/credentials/model-provider
 Credential: ../existing-credential as knowledge
 Metadata: noUserAuth: sys.model.provider.credential,knowledge
 
-#!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool load --format=markdown --flows-file=blueprint:obot --flow=obotload "ws://files/${INPUT}" "ws://files/${OUTPUT}"
+#!${GPTSCRIPT_TOOL_DIR}/bin/gptscript-go-tool load --format=markdown --flows-file=blueprint:obot --flow=obotload "ws://files/${INPUT}" -
